# docs/swagger.yaml
definitions:
  controllers.AddGameInput:
    properties:
      category_id:
        type: integer
      description:
        type: string
      genre_id:
        type: integer
      image_url:
        type: string
      name:
        type: string
    type: object
...
  controllers.GenreInput:
    properties:
      name:
        type: string
    type: object
...
  controllers.UpdateGameInput:
    properties:
      category_id:
        type: integer
      description:
        type: string
      genre_id:
        type: integer
      image_url:
        type: string
      name:
        type: string
    type: object
...
  models.Game:
    properties:
      category_id:
        type: integer
      created_at:
        type: string
      description:
        type: string
      genre_id:
        type: integer
      id:
        type: integer
      image_url:
        type: string
      name:
        type: string
      publisher_id:
        type: integer
      ratings:
        type: integer
      ratings_counter:
        type: integer
      release_date:
        type: string
      updated_at:
        type: string
    type: object
  models.Genre:
    properties:
      created_at:
        type: string
      id:
        type: integer
      name:
        type: string
      updated_at:
        type: string
    type: object
  models.InstalledGames:
    properties:
      created_at:
        type: string
      game_name:
        type: string
      id:
        type: integer
      id_game:
        type: integer
      image_url:
        type: string
      ratings:
        type: integer
      updated_at:
        type: string
      user_id:
        type: integer
    type: object
...
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
paths:
...
  /admin/add-genres:
    post:
      description: Only admin have permission to create genre
      parameters:
      - description: the body to create new genre
        in: body
        name: Body
        required: true
        schema:
          $ref: '#/definitions/controllers.GenreInput'
      - description: 'Authorization. How to input in swagger : ''Bearer <insert_your_token_here>'''
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Genre'
      security:
      - BearerToken: []
      summary: Add a new genre
      tags:
      - Admin
...
  /admin/genres/{id}:
    delete:
      description: Only admin have permission to delete genre
      parameters:
      - description: 'Authorization. How to input in swagger : ''Bearer <insert_your_token_here>'''
        in: header
        name: Authorization
        required: true
        type: string
      - description: Genre Id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerToken: []
      summary: Delete existing genre by id
      tags:
      - Admin
    patch:
      description: Only admin have permission to update genre
      parameters:
      - description: 'Authorization. How to input in swagger : ''Bearer <insert_your_token_here>'''
        in: header
        name: Authorization
        required: true
        type: string
      - description: Genre Id
        in: path
        name: id
        required: true
        type: string
      - description: the body to create new genre
        in: body
        name: Body
        required: true
        schema:
          $ref: '#/definitions/controllers.GenreInput'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Genre'
      security:
      - BearerToken: []
      summary: Update existing genre by id
      tags:
      - Admin
...
  /games/{id}:
    get:
      description: Get one game
      parameters:
      - description: Game Id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Game'
      summary: Show specific game by id
      tags:
      - Public
...
  /genres/{id}/games:
    get:
      description: Get all games of spesific genre
      parameters:
      - description: Genre Id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.Game'
            type: array
      summary: Show list of game in specific genre by genre_id
      tags:
      - Public
...
  /publisher/add-games:
    post:
      description: Only publisher have permission to publish a games
      parameters:
      - description: the body to create new games
        in: body
        name: Body
        required: true
        schema:
          $ref: '#/definitions/controllers.AddGameInput'
      - description: 'Authorization. How to input in swagger : ''Bearer <insert_your_token_here>'''
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Game'
      security:
      - BearerToken: []
      summary: Publish a new game
      tags:
      - Publisher
  /publisher/games/{id}:
    delete:
      description: Only publisher have permission to delete this game
      parameters:
      - description: 'Authorization. How to input in swagger : ''Bearer <insert_your_token_here>'''
        in: header
        name: Authorization
        required: true
        type: string
      - description: Game Id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerToken: []
      summary: Delete published game by id
      tags:
      - Publisher
    patch:
      description: Only publisher have permission to update this game
      parameters:
      - description: 'Authorization. How to input in swagger : ''Bearer <insert_your_token_here>'''
        in: header
        name: Authorization
        required: true
        type: string
      - description: Game Id
        in: path
        name: id
        required: true
        type: string
      - description: the body to create new game
        in: body
        name: Body
        required: true
        schema:
          $ref: '#/definitions/controllers.UpdateGameInput'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Game'
      security:
      - BearerToken: []
      summary: Update published game by id
      tags:
      - Publisher
...
  /users/games/{id}/install:
    patch:
      description: User installing games
      parameters:
      - description: 'Authorization. How to input in swagger : ''Bearer <insert_your_token_here>'''
        in: header
        name: Authorization
        required: true
        type: string
      - description: Game Id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.InstalledGames'
      security:
      - BearerToken: []
      summary: Install a games
      tags:
      - Users
...
  /users/installed/{id}:
    delete:
      description: Only user who have permission can uninstall this game
      parameters:
      - description: 'Authorization. How to input in swagger : ''Bearer <insert_your_token_here>'''
        in: header
        name: Authorization
        required: true
        type: string
      - description: InstalledGames Id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerToken: []
      summary: Uninstall a games
      tags:
      - Users
  /users/my-games:
    get:
      description: Get all installed games by user
      parameters:
      - description: 'Authorization. How to input in swagger : ''Bearer <insert_your_token_here>'''
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.InstalledGames'
            type: array
      security:
      - BearerToken: []
      summary: Get list of installed games by this user
      tags:
      - Users
swagger: "2.0"